+++ src/utils/mod.rs
@@ -6,5 +6,7 @@ pub mod json_schema;
 
 pub mod params;
 
+pub(crate) mod http;
+
 #[doc(hidden)]
 pub mod test_support;